add offline class extensions
