+++ src/firestore/app/offline-appkit.ts
@@ -1,0 +1,6 @@
+import { RoarAppkit } from "./appkit";
+
+export class OfflineAppKit extends RoarAppkit {
+    // TODO: find out what functions i need to override with this class
+    
+}
