+++ src/firestore/app/offline-run.ts
@@ -1,0 +1,6 @@
+import { RoarRun } from "./run";
+
+export class OfflineRun extends RoarRun {
+   // class should hold parent uid 
+// write trials to admin user's user collection instead 
+}
